requirements.txt
langchain
langchain-community
langchain-chroma
openai
faiss-cpu
python-dotenv
